audit proposal
some names changed
some functions were introduced in web3task contract
virtual header added to all
natpsec in main contract and a few in interface(lacking the remaining)
changed modifier to function to save gas on taskmanager
approval amount can be changed
mapping with reviews for keeping track of whats done
review can be triggered by both parties now like a ping pong until conclusion
creator name changed to creatorRole
errors got better
better sanity checks when calling main functions
complete task now send reward eth
